Move AmountFromValue and ValueFromAmount to rpcprotocol
Converting monetary amounts is part of the RPC protocol,
both the client and server need to share the same implementation.
