+++ src/rpcprotocol.cpp
@@ -5,6 +5,7 @@
 
 #include "rpcprotocol.h"
 
+#include "core.h"
 #include "util.h"
 
 #include <stdint.h>
@@ -250,4 +251,20 @@ Object JSONRPCError(int code, const string& message)
     error.push_back(Pair("code", code));
     error.push_back(Pair("message", message));
     return error;
+}
+
+int64_t AmountFromValue(const Value& value)
+{
+    double dAmount = value.get_real();
+    if (dAmount <= 0.0 || dAmount > 21000000.0)
+        throw JSONRPCError(RPC_TYPE_ERROR, "Invalid amount");
+    int64_t nAmount = roundint64(dAmount * COIN);
+    if (!MoneyRange(nAmount))
+        throw JSONRPCError(RPC_TYPE_ERROR, "Invalid amount");
+    return nAmount;
+}
+
+Value ValueFromAmount(int64_t amount)
+{
+    return (double)amount / (double)COIN;
 }
