+++ src/rpcserver.cpp
@@ -81,22 +81,6 @@ void RPCTypeCheck(const Object& o,
     }
 }
 
-int64_t AmountFromValue(const Value& value)
-{
-    double dAmount = value.get_real();
-    if (dAmount <= 0.0 || dAmount > 21000000.0)
-        throw JSONRPCError(RPC_TYPE_ERROR, "Invalid amount");
-    int64_t nAmount = roundint64(dAmount * COIN);
-    if (!MoneyRange(nAmount))
-        throw JSONRPCError(RPC_TYPE_ERROR, "Invalid amount");
-    return nAmount;
-}
-
-Value ValueFromAmount(int64_t amount)
-{
-    return (double)amount / (double)COIN;
-}
-
 std::string HexBits(unsigned int nBits)
 {
     union {
